requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
requests
python-dotenv
